* '247_CRM_Integration_HF_1.41.0' of bitbucket.org:anuj_aggarwal/247around-adminp-aws:
  appliance brand issue fix in uploading file
  variable issue fix
  view issue fix
  pincode file issue fixed
  brackets view changes and show defective parts in sf invoice
  minor changes in views and send mail to rm for pincode not available booking

+++ application/models/partner_model.php
@@ -147,8 +147,10 @@ class Partner_model extends CI_Model {
       return $query->result_array();
     }
 
-    function insert_data_in_batch($table_name, $rows){
-        $this->db->truncate($table_name);
+    function insert_data_in_batch($table_name, $rows,$flag=""){
+        if($flag == ""){
+            $this->db->truncate($table_name);
+        }
         return $this->db->insert_batch($table_name, $rows);
     }
